refine Segment abstraction
