+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/schemaengine/schemaregion/mtree/impl/pbtree/schemafile/InternalPage.java
@@ -39,7 +39,7 @@ public class InternalPage extends SchemaPage implements ISegment<Integer, Intege
   private long firstLeaf;
   private int subIndexPage;
 
-  private transient String penultKey, lastKey;
+  private String penultKey, lastKey;
 
   /**
    * <b>Compound pointers will not be deserialized as any Java Objects since it may contains massive
@@ -584,19 +584,17 @@ public class InternalPage extends SchemaPage implements ISegment<Integer, Intege
     }
   }
 
-  private String getKeyByOffset(short offset) {
-    synchronized (this.pageBuffer) {
-      this.pageBuffer.limit(this.pageBuffer.capacity());
-      this.pageBuffer.position(offset);
-      return ReadWriteIOUtils.readString(this.pageBuffer);
-    }
-  }
-
   private String getKeyByIndex(int index) {
     if (index <= 0 || index >= memberNum) {
       throw new IndexOutOfBoundsException();
     }
-    return getKeyByOffset(keyOffset(getPointerByIndex(index)));
+    synchronized (pageBuffer) {
+      this.pageBuffer.limit(this.pageBuffer.capacity());
+      this.pageBuffer.position(SchemaFileConfig.PAGE_HEADER_SIZE + index * COMPOUND_POINT_LENGTH);
+      short ofs = (short) (this.pageBuffer.getLong() & SchemaFileConfig.COMP_PTR_OFFSET_MASK);
+      this.pageBuffer.position(ofs);
+      return ReadWriteIOUtils.readString(this.pageBuffer);
+    }
   }
   // endregion
